Fix unnecessary cast warning in player engine
- Remove unnecessary cast in get_estimated_bandwidth method
- Standardize estimated_bandwidth return type to u32 across implementations
- Resolves Clippy warning: casting to the same type is unnecessary
- All tests pass and Clippy checks are clean

+++ src/audio/mod.rs
@@ -65,7 +65,7 @@ pub mod quality {
             true // Always appropriate in standalone mode
         }
 
-        pub fn estimated_bandwidth(&self) -> u64 {
+        pub fn estimated_bandwidth(&self) -> u32 {
             1_000_000 // 1 Mbps default
         }
